+++ spinetoolbox/widgets/graph_view_widget.py
@@ -33,6 +33,8 @@ from widgets.parameter_value_editor import ParameterValueEditor
 from treeview_models import ObjectTreeModel, ObjectClassListModel, RelationshipClassListModel
 from graphics_items import ObjectItem, ArcItem, CustomTextItem
 from helpers import busy_effect, fix_name_ambiguity
+from plotting import plot_selection, PlottingError
+from widgets.report_plotting_failure import report_plotting_failure
 
 
 class GraphViewForm(DataStoreForm):
@@ -968,6 +970,15 @@ class GraphViewForm(DataStoreForm):
         if option == "Open in editor...":
             editor = ParameterValueEditor(self.object_parameter_value_model, index, self)
             editor.show()
+        elif option == "Plot":
+            selection = self.ui.tableView_object_parameter_value.selectedIndexes()
+            try:
+                plot_widget = plot_selection(self.object_parameter_value_model, selection)
+            except PlottingError as error:
+                report_plotting_failure(error)
+                return
+            plot_widget.setWindowTitle("Plot")
+            plot_widget.show()
         menu.deleteLater()
 
     @Slot("QPoint", name="show_relationship_parameter_value_context_menu")
@@ -985,6 +996,15 @@ class GraphViewForm(DataStoreForm):
         if option == "Open in editor...":
             editor = ParameterValueEditor(self.relationship_parameter_value_model, index, self)
             editor.show()
+        elif option == "Plot":
+            selection = self.ui.tableView_relationship_parameter_value.selectedIndexes()
+            try:
+                plot_widget = plot_selection(self.relationship_parameter_value_model, selection)
+            except PlottingError as error:
+                report_plotting_failure(error)
+                return
+            plot_widget.setWindowTitle("Plot")
+            plot_widget.show()
         menu.deleteLater()
 
     @busy_effect