+++ CHANGELOG.md
@@ -31,6 +31,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - New project item: Data Interface
 - Parameter and relationship parameter values can now be edited in a dedicated editor window in Tree, Graph and Tabular
   views. The editor is accessible from when right clicking a value and selecting `Open in editor...`.
+- It is now possible to plot parameter and relationship parameter values in Tree, Graph and Tabular views.
 
 ### Changed
 - Selecting the Julia environment in Settings now requires picking the Julia interpreter **file**